--- layout: haddock ---

ForSyDe.Atom.ExB.Absent

Description

This module implements the constructors and assocuated utilities of a type which extends the behavior of a function to express "absent events" (see [Halbwachs91]).

The AbstExt type can be used directly with the atom patterns defined in ForSyDe.Atom.ExB, and no helpers or utilities are needed. Example usage:

>>> res21 (+) (Prst 1) (Prst 2)
3
>>> res21 (+) Abst     Abst
⟂
>>> filter Abst         (Prst 1)
⟂
>>> filter (Prst False) (Prst 1)
⟂
>>> filter (Prst True)  (Prst 1)
1
>>> filter' False 1 :: AbstExt Int
⟂
>>> filter' True  1 :: AbstExt Int
1
>>> degrade 0 (Prst 1)
1
>>> degrade 0 Abst
0
>>> ignore11 (+) 1 (Prst 1)
2
>>> ignore11 (+) 1 Abst
1

Incorrect usage (not covered by doctest):

λ> res21 (+) (Prst 1) Abst 
*** Exception: [ExB.Absent] Illegal occurrence of an absent and present event

Documentation

data AbstExt a

The AbstExt type extends the base type with the '⊥' symbol denoting the absence of a value/event (see [Halbwachs91]).

Constructors

Abst
⊥ denotes the absence of a value

Prst a
⊤ a present event with a value