ForSyDe-Atom

A Layered Framework for Modeling Cyber-Physical Systems

Download this project as a .zip file Download this project as a tar.gz file

ForSyDe.Atom.MoC.Time

Description

Collection of utility functions for working with Time. While the CT MoC describes time as being a non-disjoint continuum (represented in ForSyDe-Atom with Rational numbers), most of the functions here are non-ideal approximations or conversions from floating point equivalents. The trigonometric functions are imported from the numbers package, with a fixed eps parameter.

These utilities are meant to get started with using the CT MoC, and should be used with caution if result fidelity is a requirement. In this case the user should find a native Rational implementation for a particular function.

Documentation

type Time = Rational

Type alias for the type to represent metric (continuous) time. Underneath we use Rational that is able to represent any t between t₁ < t₂ ∈ T.

time :: TimeStamp -> Time

Converts TimeStamp into Time representation.

const :: a -> Time -> a

Returns a constant function.

e :: Time

Euler's number in Time format. Converted from the Prelude equivalent, which is Floating.

(*^*) :: Time -> Time -> Time

"Power of" function taking Times as arguments. Converts back and forth to Floating, as it uses the ** operator, so it is prone to conversion errors.

pi :: Time

Time representation of the number π. Rational representation with a precision of 0.000001.

sin :: Time -> Time

Sine of Time. Rational representation with a precision of 0.000001.

cos :: Time -> Time

Cosine of Time. Rational representation with a precision of 0.000001.

tan :: Time -> Time

Tangent of Time. Rational representation with a precision of 0.000001.

atan :: Time -> Time

Arctangent of Time. Rational representation with a precision of 0.000001.

asin :: Time -> Time

Arcsine of Time. Rational representation with a precision of 0.000001.

acos :: Time -> Time

Arccosine of Time. Rational representation with a precision of 0.000001.

sqrt :: Time -> Time

Square root of Time. Rational representation with a precision of 0.000001.

exp :: Time -> Time

Exponent of Time. Rational representation with a precision of 0.000001.

cosh :: Time -> Time

Hyperbolic cosine of Time. Rational representation with a precision of 0.000001.

sinh :: Time -> Time

Hyperbolic sine of Time. Rational representation with a precision of 0.000001.

tanh :: Time -> Time

Hyperbolic tangent of Time. Rational representation with a precision of 0.000001.

atanh :: Time -> Time

Hyperbolic arctangent of Time. Rational representation with a precision of 0.000001.

asinh :: Time -> Time

Hyperbolic arcsine of Time. Rational representation with a precision of 0.000001.

acosh :: Time -> Time

Hyperbolic arccosine of Time. Rational representation with a precision of 0.000001.

log :: Time -> Time

Natural logarithm of Time. Rational representation with a precision of 0.000001.

Back to top