ForSyDe.Atom.Skel.FastVector

Description

This library is is an un-official alternative to Vector meant for simulations of large data which is likely to become too cumbersome. Fast Vector functions do not use atoms, but rather use Prelude functions on a wrapped newtype using a native Haskell type (in this case lists). The API tries to copy the exported functions of ForSyDe.Atom.Skel.Vector and its submodule so that switching betwen libraries can be made seamlessly just by Vector with FastVector in the library import.

Useful links:

ForSyDe.Atom contains general guidelines for using the API

ForSyDe.Atom.Skel.Vector documents the API which this module is trying to replicate.

ForSyDe.Atom.Skel.FastVector.Matrix contains a collection of patterns for working with 2-dimensional Vectors.

ForSyDe.Atom.Skel.FastVector.Cube contains a collection of patterns for working with 3-dimensional Vectors.

ForSyDe.Atom.Skel.FastVector.DSP contains a collection of patterns commonly used in signal processing designs.

the naming convention rules on how to interpret the function names based on their number of inputs and outputs.

Documentation

newtype Vector a

In this library Vector is just a wrapper around a list.

Constructors

Vector

Fields

fromVector :: [a]

vector :: [a] -> Vector a

farm11 :: Functor f => (a -> b) -> f a -> f b

farm21 :: Applicative f => (a1 -> a2 -> b) -> f a1 -> f a2 -> f b

farm31 :: Applicative f => (a1 -> a2 -> a3 -> b) -> f a1 -> f a2 -> f a3 -> f b

farm41 :: Applicative f => (a1 -> a2 -> a3 -> a4 -> b) -> f a1 -> f a2 -> f a3 -> f a4 -> f b

farm51 :: Applicative f => (a1 -> a2 -> a3 -> a4 -> a5 -> b) -> f a1 -> f a2 -> f a3 -> f a4 -> f a5 -> f b

farm12 :: Functor f => (a -> (a1, b)) -> f a -> (f a1, f b)

farm22 :: Applicative f => (a1 -> a2 -> (a3, b)) -> f a1 -> f a2 -> (f a3, f b)

(<++>) :: Vector a -> Vector a -> Vector a infixr 5

reduce :: (t -> t -> t) -> Vector t -> t

See reduce.

length :: Vector a -> Int

See length.

drop :: Int -> Vector a -> Vector a

See drop.

take :: Int -> Vector a -> Vector a

See take.

first :: Vector t -> t

See first.

group :: Int -> Vector a -> Vector (Vector a)

See group.

fanout :: a -> Vector a

See fanout.

fanoutn :: Int -> a -> Vector a

See fanoutn.

stencil :: Int -> Vector a -> Vector (Vector a)

See stencil.

tails :: Vector a -> Vector (Vector a)

See tails.

concat :: Vector (Vector a) -> Vector a

See concat.

iterate :: Int -> (a -> a) -> a -> Vector a

See iterate.

pipe :: Vector (a -> a) -> a -> a

See pipe.

pipe1 :: (a -> t -> t) -> Vector a -> t -> t

See pipe1.

reverse :: Vector a -> Vector a

See reverse.

recuri :: Vector (b -> b) -> b -> Vector b

See recuri.

get :: Int -> Vector a -> Maybe a

See get.

evens :: Vector a -> Vector a

See evens.

odds :: Vector a -> Vector a

See odds.