ForSyDe.Atom.Skel.FastVector.DSP

Documentation

taylor

Arguments

:: (Eq a, Floating a)

=> Int
number of points in the output window.

-> Int
Number of nearly constant level sidelobes adjacent to the mainlobe

-> a
Desired peak sidelobe level in decibels (db) relative to the mainlobe

-> Vector a
The window, with the center value normalized to one (the value one appears only if the number of samples is odd).

See taylor.

taylor' :: (Eq a, Floating a) => p -> a -> Vector a

See taylor'.

dotvm :: Num a => Vector a -> Vector (Vector a) -> Vector a

See dotvm.

dotvv :: Num a => Vector a -> Vector a -> a

See dotvv.

dotvv'

Arguments

:: Num a

=> ((a -> a -> a) -> f a -> f a -> f a)
higher-order function that can wrap the (*) operation.

-> Vector (f a)

-> Vector (f a)

-> f a

See dotvv'.

dotmv'

Arguments

:: (a -> a -> a)
kernel function for a row/column reduction, e.g. (+) for dot product

-> (b -> a -> a)
binary operation for pair-wise elements, e.g. (*) for dot product

-> Vector (Vector b)
size = (xa,ya)

-> Vector a
length = xa

-> Vector a
length = xa

See dotmv'.

dotvm'

Arguments

:: (b -> b -> b)
kernel function for a row/column reduction, e.g. (+) for dot product

-> (a -> b -> b)
binary operation for pair-wise elements, e.g. (*) for dot product

-> Vector a
length = xa

-> Vector (Vector b)
size = (xa,ya)

-> Vector b
length = ya

See dotvm'.

hanning

Arguments

:: Floating n

=> Int
The length of the window

-> Vector n

See hanning.

hamming

Arguments

:: Floating n

=> Int
The length of the window

-> Vector n

See hamming.

blackman

Arguments

:: Floating n

=> Int
The length of the window

-> Vector n

See blackman.

fir

Arguments

:: Num a

=> Vector a
vector of coefficients

-> Vector a
input vector of numbers; size = n

-> Vector a
output vector of numbers; size = n

See fir.

fir'

Arguments

:: (a -> a -> a)
process/operation replacing +

-> (c -> a -> a)
process/operation replacing *

-> (a -> a)
delay process

-> Vector c
vector of coefficients

-> a
input signal/structure

-> a
output signal/structure

See fir'.

twiddles :: Floating a => Int -> Vector (Complex a)

See twiddles.

fft :: RealFloat a => Int -> Vector (Complex a) -> Vector (Complex a)

See fft.

fft' :: Floating a => (Complex a -> a -> a -> (a, a)) -> Int -> Vector a -> Vector a

See fft'.

duals :: Vector a -> (Vector a, Vector a)

See duals.

unduals :: Vector a -> Vector a -> Vector a

See unduals.

bitrev :: Vector a2 -> Vector a2

See bitrev.