ForSyDe.Atom.Skel.FastVector.Matrix

Documentation

type Matrix a = Vector (Vector a)

Matrix is simply a type synonym for vector of vectors. This means that any function on Vector works also on Matrix.

pretty

Arguments

:: Show a

=> String
separator string

-> Matrix a
input matrix

-> IO ()

Prints out to the terminal a matrix in a readable format, where all elements are right-aligned and separated by a custom separator.

>>> let m = matrix 3 3 [1,2,3,3,100,4,12,32,67]
>>> pretty "|" m
 1|  2| 3
 3|100| 4
12| 32|67

isNull :: Matrix a -> Bool

See isNull.

size :: Matrix a -> (Int, Int)

See size.

wellFormed :: Matrix a -> Matrix a

See wellFormed.

groupEvery :: Int -> [a] -> [[a]]

matrix

Arguments

:: Int
number of columns (X dimension) = x

-> Int
number of rows (Y dimension) = y

-> [a]
list of values; length = x * y

-> Matrix a
Matrix of values; size = (x,y)

See Matrix.

fromMatrix

Arguments

:: Matrix a
size = (x,y)

-> [a]
length = x * y

See fromMatrix.

fanout :: a -> Matrix a

See fanout.

indexes :: Matrix (Int, Int)

See indexes.

farm11

Arguments

:: (a -> b)

-> Matrix a
size = (xa,ya)

-> Matrix b
size = (xa,ya)

See farm11.

farm21

Arguments

:: (a -> b -> c)

-> Matrix a
size = (xa,ya)

-> Matrix b
size = (xb,yb)

-> Matrix c
size = (minimum [xa,xb], minimum [ya,yb])

See farm21.

farm31

Arguments

:: (a -> b -> c -> d)

-> Matrix a
size = (xa,ya)

-> Matrix b
size = (xb,yb)

-> Matrix c
size = (xc,yc)

-> Matrix d
size = (minimum [xa,xb,xc], minimum [ya,yb,yc])

See farm31.

reduce :: (a -> a -> a) -> Matrix a -> a

See reduce.

dotV

Arguments

:: (a -> a -> a)
kernel function for a row/column reduction, e.g. (+) for dot product

-> (b -> a -> a)
binary operation for pair-wise elements, e.g. (*) for dot product

-> Matrix b
size = (xa,ya)

-> Vector a
length = xa

-> Vector a
length = xa

See fotV.

dot

Arguments

:: (a -> a -> a)
kernel function for a row/column reduction, e.g. (+) for dot product

-> (b -> a -> a)
binary operation for pair-wise elements, e.g. (*) for dot product

-> Matrix b
size = (xa,ya)

-> Matrix a
size = (ya,xa)

-> Matrix a
size = (xa,xa)

See dot.

get

Arguments

:: Int
X index starting from zero

-> Int
Y index starting from zero

-> Matrix a

-> Maybe a

See get.

take

Arguments

:: Int
X index starting from zero

-> Int
Y index starting from zero

-> Matrix a

-> Matrix a

See take.

drop

Arguments

:: Int
X index starting from zero

-> Int
Y index starting from zero

-> Matrix a

-> Matrix a

See drop.

crop

Arguments

:: Int
crop width = w

-> Int
crop height = h

-> Int
X start position = x0

-> Int
Y start position = y0

-> Matrix a
size = (xa,ya)

-> Matrix a
size = (minimum [w,xa-x0], minimum [h,xa-x0])

See crop.

group

Arguments

:: Int
width of groups = w

-> Int
height of groups = h

-> Matrix a
size = (xa,ya)

-> Matrix (Matrix a)
size = (xa div w,ya div h)

See group.

stencil :: Int -> Int -> Matrix a -> Matrix (Matrix a)

See stencil.

reverse :: Matrix a -> Matrix a

See reverse.

transpose

Arguments

:: Matrix a
X:Y orientation

-> Matrix a
Y:X orientation

See transpose.

replace :: Int -> Int -> Matrix a -> Matrix a -> Matrix a

See replace.