ForSyDe.Atom.MoC.Stream

Description

This module defines the shallow-embedded Stream data type and utility functions operating on it. In ForSyDe a signal is represented as a (partially or totally) ordered sequence of events that enables processes to communicate and synchronize. In ForSyDe-Atom signals are the main operating type for the MoC layer. The Stream type is only the base (potentially infinite) structure which can encapsulate events in order to describe signals.

Documentation

data Stream e

Defines a stream of events, encapsulating them in a structure isomorphic to an infinite list [Bird87], thus all properties of lists may also be applied to Streams. While, in combination with lazy evaluation, it is possible to create and simulate infinite signals, we need to ensure that the first/previous event is always fully evaluated, which is equivalent to ensuring the the monotonicity property in dataflow systems. This translates in the following composition rule:

non-causal feedback is forbidden

also called "zero-delay" feedback loops, are caused by un-evaluated self-referential calls. In a feedback loop, there always has to be enough events to ensure the data flow.

This rule imposes that the stream of data is uninterrupted in order to have an evaluatable kernel every time a new event is produced (i.e. to avoid deadlocks), which is ther equivalent to ensuring continuity in dataflow systems. This translates in the following rule:

cleaning of signals in a feedback is forbidden

in other words, whenever a feedback composition occurs, for each new input at any instant in time, a process must react with at least one output event.

Constructors

NullS
terminates a signal

e :- (Stream e) infixr 3
the default constructor appends an event to the head of the stream

stream :: [a] -> Stream a

The function stream converts a list into a stream.

fromStream :: Stream a -> [a]

The function fromStream converts a signal into a list.

headS :: Stream a -> a

Returns the head of a signal.

tailS :: Stream e -> Stream e

Returns the tail of a signal

lastS :: Stream p -> p

Returns the last event in a signal.

repeatS :: a -> Stream a

Returns an infinite list containing the same repeated event.

takeS :: (Num t, Ord t) => t -> Stream e -> Stream e

Returns the first n events in a signal.

dropS :: (Num t, Ord t) => t -> Stream e -> Stream e

Drops the first n events in a signal.

takeWhileS :: (a -> Bool) -> Stream a -> Stream a

Returns the first events of a signal which comply to a condition.

(+-+) :: Stream e -> Stream e -> Stream e

Concatenates two signals.