--- layout: haddock ---

ForSyDe.Atom.MoC.Stream

Description

This module defines the shallow-embedded Stream datatype and utility functions operating on it. In ForSyDe a signal is represented as a (partially or totally) ordered sequence of events that enables processes to communicate and synchronize. The Stream type is but an ordered structure to encapsulate events as infinite streams.

Documentation

data Stream e

Defines a stream of events, encapsulating them in a structure isomorphic to an infinite list [Bird87], thus all properties of lists may also be applied to Streams. While, in combination with lazy evaluation, it is possible to create and simulate infinite signals, we need to ensure that the first/previous event is always fully evaluated. This can be translated into the following rule:

zero-delay feedbacks

are forbidden, due to un-evaluated self-referential calls. In a feedback loop, there always has to be enough events to ensure the data flow.

This rule imposes that the stream of data is uninterrupted in order to have an evaluatable kernel every time a new event is produced (i.e. to avoid deadlocks). Thus we can add the rule:

cleaning of output events

is also forbidden. In other words, for each new input at any instant in time, a process must react with at least one output event.

Constructors

NullS
terminates a signal

e :- (Stream e) infixr 3
the default constructor appends an event to the head of the stream

stream :: [a] -> Stream a

The function signal converts a list into a signal.

fromStream :: Stream a -> [a]

The function fromStream converts a signal into a list.

headS :: Stream a -> a

Returns the head of a signal.

tailS :: Stream e -> Stream e

Returns the tail of a signal

lastS :: Stream p -> p

Returns the last event in a signal.

repeatS :: a -> Stream a

Returns an infinite list containing the same repeated event.

takeS :: (Ord t, Num t) => t -> Stream e -> Stream e

Returns the first n events in a signal.

dropS :: (Ord t, Num t) => t -> Stream e -> Stream e

Drops the first n events in a signal.

takeWhileS :: (a -> Bool) -> Stream a -> Stream a

Returns the first events of a signal which comply to a condition.

(+-+) :: Stream e -> Stream e -> Stream e

Concatenates two signals.