--- layout: haddock ---

ForSyDe.Atom.MoC.TimeStamp

Description

This module implements a timestamp data type, based on Data.Time.Clock.

Documentation

type TimeStamp = DiffTime

Alias for the type representing discrete time. It is inherently quantizable, the quantum being a picosecond (10⁻¹² seconds), thus it can be considered order-isomorphic with a set of integers, i.e. between any two timestamps there is a finite number of timestamps. Moreover, a timestamp can be easily translated into a rational number representing fractions of a second, so the conversion between timestamps (discrete time) and rationals (analog/continuous time) is straightforward.

This type is used in the explicit tags of the DE MoC (and subsequently the discrete event evaluation engine for simulating the CT MoC).

picosec :: Integer -> TimeStamp

Specifies a timestamp in terms of picoseconds.

nanosec :: Integer -> TimeStamp

Specifies a timestamp in terms of nanoseconds.

microsec :: Integer -> TimeStamp

Specifies a timestamp in terms of microseconds.

milisec :: Integer -> TimeStamp

Specifies a timestamp in terms of miliseconds.

sec :: Integer -> TimeStamp

Specifies a timestamp in terms of seconds.

minutes :: Integer -> TimeStamp

Specifies a timestamp in terms of minutes.

hours :: Integer -> TimeStamp

Specifies a timestamp in terms of hours.

toTime :: TimeStamp -> Rational

Converts a timestamp to a rational number, used for describing continuous time.

pi :: TimeStamp

TimeStamp representation of the number π. Converted from the Prelude equivalent, which is Floating.

Orphan instances

Read DiffTime
reads [n]s as a fraction of a second, where [n] is a floating point number.

Instance details

Methods

readsPrec :: Int -> ReadS DiffTime

readList :: ReadS [DiffTime]

readPrec :: ReadPrec DiffTime

readListPrec :: ReadPrec [DiffTime]