--- layout: haddock ---

ForSyDe.Atom.Utility.Plot

Description

This module imports plotting and data dumping functions working with "plottable" data types, i.e. instances of the Plot and Plottable type classes.

User API

The following commands are frequently used as part of the normal modeling routine.

Configuration settings

data Config

Record structure containing configuration settings for the plotting commands.

Constructors

Cfg

Fields

verbose :: Bool

verbose printouts on terminal

path :: String

directory where all dumped files will be found

title :: String

base name for dumped files

rate :: Float

sample rate if relevant. Useful for explicit-tagged signals, ignored otherwise.

xmax :: Float

Maximum X coordinate. Mandatory for infinite structures, optional otherwise.

labels :: [String]

list of labels with the names of the structures plotted

fire :: Bool

if relevant, fires a plotting or compiling program.

other :: Bool

if relevant, dumps additional scripts and plots.

defaultCfg :: Config

Default configuration: verbose, dump everything possible, fire whatever program needed. Check source for settings.

Example usage:

>>> defaultCfg {xmax = 15, verbose = False, labels = ["john","doe"]}
Cfg {verbose = False, path = "./fig", title = "plot", rate = 1.0e-2, xmax = 15.0, labels = ["john","doe"], fire = True, other = True}

silentCfg :: Config

Silent configuration: does not fire any program or print our unnecessary info. Check source for settings.

noJunkCfg :: Config

Clean configuration: verbose, does not dump more than necessary, fire whatever program needed. Check source for settings.

Data preparation

prepare

Arguments

:: Plot a

=> Config
configuration settings

-> a
plottable data type

-> PlotData
structure ready for dumping

Prepares a single plottable data structure to be dumped and/or plotted.

prepareL :: Plot a => Config -> [a] -> PlotData

Prepares a list of plottable data structures to be dumped and/or plotted. See prepare.

prepareV :: Plot a => Config -> Vector a -> PlotData

Prepares a vector of plottable data structures to be dumped and/or plotted. See prepare.

Dumping and plotting data

showDat :: PlotData -> IO ()

Prints out the sampled contents of a prepared data set.

dumpDat :: PlotData -> IO [String]

Dumps the sampled contents of a prepared data set into separate .dat files.

plotGnu :: PlotData -> IO ()

Generates a GNUplot script and .dat files for plotting the sampled contents of a prepared data set. Depending on the configuration settings, it also dumps LaTeX and PDF plots, and fires the script.

OBS: needless to say that GNUplot needs to be installed in order to use this command. Also, in order to fire GNUplot from a ghci session you might need to install gnuplot-x11.

heatmapGnu :: PlotData -> IO ()

Similar to plotGnu but creates a heatmap plot using the GNUplot engine. For this, the input needs to contain at least two columns of data, otherwise the plot does not show anything, i.e. the samples need to be lists or vectors of two or more elements.

OBS: same dependencies are needed as for plotGnu.

showLatex :: PlotData -> IO ()

Prints out a LaTeX environment from a prepared data set. This environment should be paste inside a tikzpicture in a document title which imports the ForSyDe-LaTeX package.

dumpLatex :: PlotData -> IO [String]

Dumps a set of formatted data files with the extension .flx that can be imported by a LaTeX document which uses the ForSyDe-LaTeX package.

plotLatex :: PlotData -> IO ()

Creates a standalone LaTeX document which uses the ForSyDe-LaTeX package, plotting a prepared data set. Depending on the configuration settings, the command pdflatex may also be invoked to compile a pdf image.

OBS: A LaTeX compiler is required to run the pdflatex command. The ForSyDe-LaTeX package also needs to be installed according to the instructions on the project web page.

The data types

Below the data types involved are shown and the plottable structures are documented.

class Plottable a where

This class gathers types which can be sampled and converted to a numerical string which can be read and interpreted by a plotter engine.

Minimal complete definition

toCoord

Methods

toCoord :: a -> String

Transforms the input type into a coordinate string.

class Plot a where

This class gathers all ForSyDe-Atom structures that can be plotted.

Minimal complete definition

(sample | sample'), takeUntil, getInfo

Methods

sample :: Float -> a -> Samples

Samples the data according to a given step size.

sample' :: a -> Samples

Samples the data according to the internal structure.

takeUntil :: Float -> a -> a

Takes the first samples until a given tag.

getInfo :: a -> PInfo

Returns static information about the data type.

data PInfo

Static information of each plottable data type.

Constructors

Info

Fields

typeid :: String

id used usually in implicit tags

command :: String

LaTeX identifier

measure :: String

unit of measure

style :: String

style tweaking in the GNUplot script

stacking :: Bool

if the plot is stacking

sparse :: Bool

if the sampled data is sparse instead of dense

type Samples = [(String, String)]

Alias for sampled data

type PlotData = (Config, PInfo, [(String, Samples)])

Alias for a data set prepared to be plotted.